feat: limit kitchen staff view
