+++ src/app/recipe/page.tsx
@@ -1,7 +1,5 @@
-"use client";
-
 import { CategoryValue, EMPTY_FILTERS } from "@/lib/types";
-import { useState } from "react";
+// import { useState } from "react";
 import { useMealData } from "@/hooks/useMealData";
 import { Recipe, Combo } from "@/lib/types";
 import { Suspense } from "react";
@@ -9,11 +7,19 @@ import MealBrowser from "@/components/MealBrowser";
 import FilterMenu from "@/components/FilterMenu";
 import ViewRecipePopUp from "@/components/ViewRecipePopUp";
 import RecipePageClient from "@/components/RecipePageClient";
+import { auth } from "@clerk/nextjs/server";
+import dbConnect from "@/database/db";
+import User from "@/database/UserSchema";
 
-export default function RecipePage() {
+export default async function RecipePage() {
+  const { userId } = await auth();
+
+  await dbConnect();
+  const currentUser = await User.findOne({ clerkId: userId });
+
   return (
     <Suspense fallback={<div>Loading...</div>}>
-      <RecipePageClient />
+      <RecipePageClient userRole={currentUser.role} />
     </Suspense>
   );
 }
