+++ src/components/CardGrid.tsx
@@ -16,6 +16,7 @@ type Props = {
   selectedIds?: Set<string>;
   onToggleSelect?: (id: string, name: string) => void;
   onOpenItem?: (item: Recipe | Combo) => void;
+  userRole?: string;
 };
 
 export default function CardGrid({
@@ -28,6 +29,7 @@ export default function CardGrid({
   selectedIds,
   onToggleSelect,
   onOpenItem,
+  userRole,
 }: Props) {
   if (loading) return <div className="text-sm text-black/60">Loading…</div>;
   if (error) return <div className="text-sm text-red-600">{error}</div>;
@@ -54,6 +56,7 @@ export default function CardGrid({
             isSelected={selectedIds?.has(combo._id)}
             onSelect={() => onToggleSelect?.(combo._id, combo.name)}
             onOpen={() => onOpenItem?.(combo)}
+            userRole={userRole}
           />
         ))}
       </div>
@@ -79,6 +82,7 @@ export default function CardGrid({
           isSelected={selectedIds?.has(recipe._id)}
           onSelect={() => onToggleSelect?.(recipe._id, recipe.name)}
           onOpen={() => onOpenItem?.(recipe)}
+          userRole={userRole}
         />
       ))}
     </div>
